pyproject.toml
[project]
name = "sports-bet"
version = "0.1.0"
description = "スポーツベット ハンデ予測システム"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "pandas>=2.1",
    "numpy>=1.24",
    "sqlalchemy>=2.0",
    "lightgbm>=4.0",
    "scikit-learn>=1.3",
    "typer>=0.9",
    "streamlit>=1.30",
    "pyyaml>=6.0",
    "rich>=13.0",
    "matplotlib>=3.8",
    "plotly>=5.18",
]

[project.scripts]
sports-bet = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
